Convert elf2tab into a lib+bin crate
The rationale for this change is to allow cargo-raze to process this
crate as a dependency.  As a bin-only crate, cargo-raze requires
the crate to be published at crates.io and does not permit using a
unofficial git repo as the source.

As a bin+lib crate, there is no such restriction.
